use findAddressCandidates consistently
resolves #107

this change makes consistent use of the findAddressCandidates operation of the World Geocoding Service.  this makes it unnecessary to differentiate and maintain simple() and addresses() methods for geocoding.

wrote out a few more samples in the markdown doc to demonstrate how to pass additional parameters and make requests to custom services.

because the response signature of findAddressCandidates differs from find, my hunch was to group the additional cleanup and call this a breaking change.